Removed old code, made some methods in FrequencyChart static. Added text. Made test for non-gui part.
